Major changes to energy-/resistance-/weaknessType
WeaknessType: Made all the properties in a private visibility, changed the construct to only energytype and multiplier and removed the values for the multipliers from the page.

ResistanceType: Made all the properties in a private visibility, changed the construct to only energytype and worth and removed the values from the page

EnergyType: Removed all properties and left it to only energyType property in a private visibility. Changed the construction to only include energytype and created the energytypes in the index as classes. Removed the values in this page