Changes to properties, constructor, getter/setter & pikachu/charmeleon classes
Made all the properties protected visibility.
Added the proper comments to the code
Included a worth & multiplier property 
Made getter & setter methods to use the right values of the properties
Proper creation of pikachu & charmeleon classes and the extending of the parent constructor with the base values for those pokemons.