Test echos, annotation, if statement calculations & function
Created a class & function for the battle WIP as it cannot be called yet.

Put some test echos on top to see whether the page can read all the given properties completely, it cannot at this time.

Changed the annotation to properly fit the page with the given rules on how it should be made.

Created a commented section for if statement that handles the calculations, hard WIP as there's no idea yet on how to handle this if statement to fill in what it has to.