Test echos/comments/calling/function
Removed the test echos after the information it needed to echo was showing

added comments on who is who for pokemonDefender & pokemonAttacker

Changed every object name to the parameters pokemonDefender & pokemonAttacker

Put back in the calculations, the unique ones havent changed and the repeated calculations are put into a method and called in the code.

Removed the if comment statements I had made as they were wrong.